every class to its own file + dir objs
