+++ load_and_post.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 #-*- coding: UTF 8 -*-
 
-from bot import ExportBot
+from objs.bot import ExportBot
 bot = ExportBot()
 bot.detect()
 bot.public_posts()
