+++ objs/bot.py
@@ -1,0 +1,80 @@
+# -*- coding: UTF 8 -*-
+# !/usr/bin/env python3
+import configparser
+# import socket
+import logging
+from string import punctuation
+import telegram
+import base64
+from datetime import datetime
+
+from objs.database_handler import Database
+from objs.rss_parser import *
+from objs.bitly import Bitly
+
+
+# socket.setdefaulttimeout(10)
+
+
+class ExportBot(object):
+    def __init__(self):
+        config = configparser.ConfigParser()
+        config.read('./config')
+        sources = configparser.ConfigParser()
+        sources.read('./sources')
+        tokens = configparser.ConfigParser()
+        tokens.read('./tokens')
+        log_file = config['Export_params']['log_file']
+        self.pub_pause = int(config['Export_params']['pub_pause'])
+        self.delay_between_messages = int(
+            config['Export_params']['delay_between_messages'])
+        logging.basicConfig(
+            format=u'%(filename)s[LINE:%(lineno)d]# %(levelname)-8s [%(asctime)s] %(message)s',
+            level=logging.INFO,
+            handlers=[logging.FileHandler(u'%s' % log_file, 'w', 'utf-8')])
+        self.db = Database(config['Database']['Path'])
+        self.src = Source(sources['RSS'])
+        self.chat_id = config['Telegram']['chat']
+        self.bot_access_token = tokens['Telegram']['access_token']
+        self.bot = telegram.Bot(token=self.bot_access_token)
+        self.bit_ly = Bitly(tokens['Bitly']['access_token'])
+
+    def detect(self):
+        # получаем 30 последних постов из rss-канала
+        self.src.refresh()
+        news = self.src.news
+        news.reverse()
+        # Проверяем на наличие в базе ссылки на новость, если нет, то добавляем в базу данных с
+        # отложенной публикацией
+        for i in news:
+            if not self.db.find_link(i.link):
+                now = int(time.mktime(time.localtime()))
+                i.publish = now + self.pub_pause
+                logging.info(u'Detect news: %s' % str(i))
+                self.db.add_news(i)
+
+    def public_posts(self):
+        now = datetime.now()
+        # Получаем 30 последних записей из rss канала и новости из БД, у которых message_id=0
+        posts_from_db = self.db.get_post_without_message_id()
+        line = [i for i in self.src.news if (
+            now - datetime.fromtimestamp(i.date)).days < 1]
+        # Выбор пересечний этих списков
+        for_publishing = list(set(line) & set(posts_from_db))
+        for_publishing = sorted(for_publishing, key=lambda news: news.date)
+        # for_publishing = sorted(line, key=lambda news: news.date)
+        # Постинг каждого сообщения
+        for post in tqdm(for_publishing, desc="Posting news"):
+            header = base64.b64decode(post.text).decode('utf8')
+            header = ''.join(c for c in header if c not in set(punctuation + '—«»'))
+            header = '#' + '_'.join(header.lower().split())
+            text = '%s %s' % (header,
+                              self.bit_ly.short_link(base64.b64decode(post.link).decode('utf8')))
+            a = self.bot.sendMessage(
+                chat_id=self.chat_id, text=text)  # , parse_mode=telegram.ParseMode.HTML)
+            message_id = a.message_id
+            chat_id = a['chat']['id']
+            self.db.update(post.link, chat_id, message_id)
+            logging.info(u'Public: %s;%s;' %
+                         (str(post), message_id))
+            time.sleep(self.delay_between_messages)
